Implemented libpng. Ported to Linux and partially to Android. - Textures are loaded using libpng - Screen re-implemented as singleton - Padded font glyphs for proper font anti-aliasing - Deprecated Label.h - Optimizations - Fixed memory leaks - Clean up - Still needs fixing: Touch register

+++ libs/Rainbow/FontGlyph.h
@@ -8,7 +8,7 @@
 
 struct FontGlyph
 {
-	int advance;
+	short int advance;
 	SpriteVertex quad[4];
 };
